update db and updated column solded in product table

+++ database/migrations/2020_06_05_153255_create_bill_detail_table.php
@@ -18,7 +18,7 @@ class CreateBillDetailTable extends Migration
             $table->integer('id_bill');
             $table->integer('id_product');
             $table->integer('quantity')->default(1);
-            $table->double('unit_price');
+            $table->integer('unit_price');
             $table->timestamps();
         });
     }
